changed caching method to function, include cite
cite data is now cached at the same time. the Publicaton methods don't, as of now,
check the global cache for either the result or citation, so the user is
responsible for checking if it exists. some data such as "level of
search" is not saved. current plan is to save the JSON and recreate the
objects from them, but that will lose search data