debugging NoneType global_indent in test
the indent of zero works in the call to the get_cite method, but not in
the subsequent call to the get_cited_by method. this is not due to a
difference in the methods but the intervening

core.global_indent += 2

which changes global_indent to a NoneType. changing this line to

core.global_indent = core.global_indent + 2

keeps the same error.