simple algorithms for exploring citation networks
these can be used to generate citation networks, when applying filters.
this is especially useful when one wants to investigate the citation
history of a particular publiction.

at this point there is no resolving of cross-citations, so that there is
simply a tree structure. however, even at the first level in many
citation networks there will be cross-links (citation networks tend to
be clustered)