untested (except local unit test) caching changes
the reason to have local caching is that SerpAPI only holds in its
servers a cache for some time, which gives redundant queries (unless an
update is desired).

the caching done here is not what one expects as regards cited by
queries, in particular, only the organic results are saved. this is
because by the time you do a cited by query, you already have the
article information (from the search query done to obtain the article).
but there may be some value in simply caching the json returned by the
SerpAPI.

also using the title as an effective hash for the json result may not be
optimal, as there may be similarly titled works