implemented application object 'cache' (=dict)
in the from_json constructor for a publication, if the publication
object already exists in memory, instead it is returned rather than a
new. there's no way to do this with the init method so instead a warning
is thrown. this will allow general graph connections to be made between
application objects.

the DFS and BFS algorithms will have to be modified to prevent
infinite loops, by keeping a dictionary of visited keys ('black' rather
than 'white' nodes)