can now search for specified number of results
this saves query resources in the case of diminishing returns for a
given search term. this depends on the NUM_RESULTS_PAGE not being
changed, since it isn't possible with the given flattening of organic
results to infer the number of results per page (and the search
parameter for determining that, while present in the hashed dict, isn't
in the JSON result). can be fixed by keeping data not flat