added test_logger for testing
a unit testing framework would make this redundant and anyway would have
some advantage such as automating the execution of the unit tests.

also, the methods working on the dummy data don't give sensible results
such as unique query results. some of this may be as expected since the link
finding is emulated by copying the same dictionary, but in that case the
dummy data routines should be changed