made cache directory structure
cache and load_cache requires the datatype which is to be cached/loaded
to be given. this is provided as a string, since what is cached is
anyway the result of a string query parameter (if the objects were being
effectively serialized, the class could be given as the argument). the
publication caching is redundant because the information is in the
search (unlike the cite information from following the link)