made reqget stateful (global server class instance)
this ensures that every returned organic result has a unique title, and
therefore every publication generaed from a given search result is
unique. the test-publication and test-search assertions for the
uniqueness of returned data still fail.

makes use of the copy.deepcopy function
because the json-type data are nested dictionaries for which there has
to be a recursive copying.