remove characterize function
replace with a characterize_directed. this 'characterize' would be in
alalogy 'describe' for relational database tables, but that is done by
the networks 101. and anyway the properties for a graph are much greater
making it cost too much to really enumerate them.  a series of data is
usually only characterized with the datatype, number of distinct values
(if not infinite), and its moments.